+++ noodles-fasta/src/record/definition.rs
@@ -5,7 +5,7 @@ use std::{
     str::{self, FromStr},
 };
 
-use bstr::ByteSlice;
+use bstr::{BStr, BString};
 
 const PREFIX: char = '>';
 
@@ -15,8 +15,8 @@ const PREFIX: char = '>';
 /// description.
 #[derive(Clone, Debug, Eq, PartialEq)]
 pub struct Definition {
-    name: Vec<u8>,
-    description: Option<Vec<u8>>,
+    name: BString,
+    description: Option<BString>,
 }
 
 impl Definition {
@@ -28,9 +28,9 @@ impl Definition {
     /// use noodles_fasta::record::Definition;
     /// let definition = Definition::new("sq0", None);
     /// ```
-    pub fn new<N>(name: N, description: Option<Vec<u8>>) -> Self
+    pub fn new<N>(name: N, description: Option<BString>) -> Self
     where
-        N: Into<Vec<u8>>,
+        N: Into<BString>,
     {
         Self {
             name: name.into(),
@@ -47,8 +47,8 @@ impl Definition {
     /// let definition = Definition::new("sq0", None);
     /// assert_eq!(definition.name(), b"sq0");
     /// ```
-    pub fn name(&self) -> &[u8] {
-        &self.name
+    pub fn name(&self) -> &BStr {
+        self.name.as_ref()
     }
 
     /// Returns the description if it is set.
@@ -56,25 +56,26 @@ impl Definition {
     /// # Examples
     ///
     /// ```
+    /// use bstr::{BStr, BString};
     /// use noodles_fasta::record::Definition;
     ///
     /// let definition = Definition::new("sq0", None);
     /// assert_eq!(definition.description(), None);
     ///
-    /// let definition = Definition::new("sq0", Some(Vec::from("LN:13")));
-    /// assert_eq!(definition.description(), Some(&b"LN:13"[..]));
+    /// let definition = Definition::new("sq0", Some(BString::from("LN:13")));
+    /// assert_eq!(definition.description(), Some(BStr::new("LN:13")));
     /// ```
-    pub fn description(&self) -> Option<&[u8]> {
-        self.description.as_deref()
+    pub fn description(&self) -> Option<&BStr> {
+        self.description.as_ref().map(|s| s.as_ref())
     }
 }
 
 impl fmt::Display for Definition {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(f, "{PREFIX}{}", self.name.as_bstr())?;
+        write!(f, "{PREFIX}{}", self.name)?;
 
         if let Some(description) = self.description() {
-            write!(f, " {}", description.as_bstr())?;
+            write!(f, " {}", description)?;
         }
 
         Ok(())
@@ -137,7 +138,7 @@ mod tests {
         let definition = Definition::new("sq0", None);
         assert_eq!(definition.to_string(), ">sq0");
 
-        let definition = Definition::new("sq0", Some(Vec::from("LN:13")));
+        let definition = Definition::new("sq0", Some(BString::from("LN:13")));
         assert_eq!(definition.to_string(), ">sq0 LN:13");
     }
 
@@ -147,7 +148,7 @@ mod tests {
 
         assert_eq!(
             ">sq0  LN:13".parse(),
-            Ok(Definition::new("sq0", Some(Vec::from("LN:13"))))
+            Ok(Definition::new("sq0", Some(BString::from("LN:13"))))
         );
 
         assert_eq!("".parse::<Definition>(), Err(ParseError::Empty));